% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_list.R
\name{black_list}
\alias{black_list}
\title{black_list}
\usage{
black_list(data, cols, freq = 12, white_list = FALSE)
}
\arguments{
\item{data}{a data frame including factor variables}

\item{cols}{cols to look at minimum frequey}

\item{freq}{int miniumum frequency}

\item{white_list}{logical use a white list (inclusion) or blacklist exclusion}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
