% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{chunk_string}
\alias{chunk_string}
\title{chunk_string}
\usage{
chunk_string(x, chunk_len = 3)
}
\arguments{
\item{x}{a text string (not a vector)}

\item{chunk_len}{integer nchar length of each part of the key}
}
\value{
a vector of the string chuncks
}
\description{
splits a string in to chunks of length chunk_len, returns a vector
}
\author{
Matthew Davis
}
