% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_add}
\alias{cp_add}
\title{cp add}
\usage{
cp_add(x, n = 6)
}
\arguments{
\item{x}{int vector of calendar periods ie: 201202}

\item{n}{int number of months to add}
}
\value{
a integer calendar period
}
\description{
adds months to a calendar period, hidden helper function
}
\author{
Matthew Davis
}
