% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_add_1}
\alias{cp_add_1}
\title{cp add 1}
\usage{
cp_add_1(x, n)
}
\arguments{
\item{x}{a signle calendar period ie: 201202}

\item{n}{int number of months to add}
}
\value{
a integet calendar period
}
\description{
adds months to a calendar period, hidden helper function
}
\author{
Matthew Davis
}
