% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_seq}
\alias{cp_seq}
\title{cp_seq}
\usage{
cp_seq(x, n = 6, use_first = TRUE)
}
\arguments{
\item{x}{int calendar period to start ie 201202}

\item{n}{int number of months to add in the sequence}

\item{use_first}{logical, in cluded x (as the start of the sequence)}
}
\value{
a integer calendar period
}
\description{
creates a sequence of calendar periods
}
\author{
Matthew Davis
}
