% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_to_calendar_1}
\alias{cp_to_calendar_1}
\title{CP to Calendar}
\usage{
cp_to_calendar_1(x, fp = FALSE)
}
\arguments{
\item{x}{int calendar period to start ie 201202}

\item{fp}{whether the calindar period is fiscal}
}
\value{
a vector of business days and holiday/weekends
}
\description{
finds holidays and buisness days for a calendar period
}
\author{
Matthew Davis
}
