% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_to_calendar_np}
\alias{cp_to_calendar_np}
\title{CP to Calendar non parallel}
\usage{
cp_to_calendar_np(x, fp = FALSE)
}
\arguments{
\item{x}{int calendar period to start ie 201202}

\item{fp}{whether the calindar period is fiscal}
}
\value{
a vector of business days and holiday/weekends
}
\description{
creates a sequence of calendar periods
}
\author{
Matthew Davis
}
