% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_periods.R
\name{cp_to_fp}
\alias{cp_to_fp}
\title{cp to fp}
\usage{
cp_to_fp(x, n = 6, rev = FALSE)
}
\arguments{
\item{x}{int vector of calendar periods ie: 201202}

\item{n}{int number of months to add}

\item{rev}{logical, reverse and go from fiscal period to calendar period}
}
\value{
a integer calendar periods
}
\description{
adds months to a calendar period, hidden helper function
}
\author{
Matthew Davis
}
