% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyscale.R
\name{keyscale}
\alias{keyscale}
\title{Key Scaler}
\usage{
keyscale(data, params = list())
}
\arguments{
\item{data}{a data frame like object}

\item{params}{list "key_col = NULL, trans_cols = NULL, scale = TRUE, center = TRUE"}
}
\value{
a keyscale_obj
}
\description{
creates seperate scales by a key
}
\author{
Matthew Davis
}
