% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_coder.R
\name{level_coder}
\alias{level_coder}
\title{level_coder}
\usage{
level_coder(data)
}
\arguments{
\item{data}{a data frame including factor variables}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
