% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_cp.R
\name{missing_cp}
\alias{missing_cp}
\title{missing calendar periods}
\usage{
missing_cp(data, return_completed = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{a vector of calendar perionds, c(201201, 201202)}

\item{return_completed}{logical return only missing periods, or a completed vector of calendar periods}

\item{verbose}{print debugging}
}
\value{
a vector of missing calendar periods, or a vector of complete calendar periods from min to max data
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
