% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_list.R
\name{predict.bl}
\alias{predict.bl}
\title{predict method for blacklist object}
\usage{
\method{predict}{bl}(object, data)
}
\arguments{
\item{data}{a data frame including factor variables}

\item{cols}{cols to look at minimum frequey}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
