% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_encoder.R
\name{predict.encoder}
\alias{predict.encoder}
\title{predict method sparseEncoder}
\usage{
\method{predict}{encoder}(object, data, ...)
}
\arguments{
\item{object}{a dataframe with factor columns}

\item{data}{a dataframe to encode using pre trained encoder}

\item{...}{additional arguments}
}
\value{
a sparse matrix
}
\description{
transformation/predict method for catgories
}
