% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyscale.R
\name{predict.keyscale_obj}
\alias{predict.keyscale_obj}
\title{Predict Key Scale}
\usage{
\method{predict}{keyscale_obj}(object, data, rev = FALSE)
}
\arguments{
\item{object}{a keyscale_obj}

\item{data}{new data frame like object}

\item{rev}{unscale a scaled object}
}
\value{
a keyscale_obj
}
\description{
creates seperate scales by a key
}
\author{
Matthew Davis
}
