% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_encoder.R
\name{sparseEncoder}
\alias{sparseEncoder}
\title{sparseEncoder}
\usage{
sparseEncoder(data, minFreq = 0.01, imputer_fun = median, f = formula(~. -
  1), verbose = FALSE)
}
\arguments{
\item{data}{a dataframe with factor columns}

\item{minFreq}{minimum frequency of occucrence of factor levels}

\item{imputer_fun}{function to pass in for missing value imputation}

\item{f}{formula for encoding}

\item{verbose}{print debugging}
}
\value{
a transformation object
}
\description{
sparse one hot catagorical encoder
}
