% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{split_key}
\alias{split_key}
\title{Split Key}
\usage{
split_key(x, chunk_len = 3)
}
\arguments{
\item{x}{a character vector of text}

\item{chunk_len}{integer nchar length of each part of the key}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
