% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyscale.R
\name{keyscale}
\alias{keyscale}
\title{Key Scaler}
\usage{
keyscale(data, val_col = NULL, key_col = NULL, verbose = FALSE)
}
\arguments{
\item{data}{a data frame like object}

\item{val_col}{char vector of value columns (optional)}

\item{key_col}{key column (a seperate key will be made for each key)}

\item{verbose}{print debugging}
}
\value{
a keyscale_obj
}
\description{
creates seperate scales by a key
}
\author{
Matthew Davis
}
