% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_coder.R
\name{level_coder}
\alias{level_coder}
\title{level_coder}
\usage{
level_coder(data, verbose = TRUE, rm_null_levs = TRUE)
}
\arguments{
\item{data}{a data frame including factor variables}

\item{verbose}{print debugging}

\item{rm_null_levs}{logical whether to remove null levels}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
