% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_list.R
\name{predict.bl}
\alias{predict.bl}
\title{predict method for blacklist object}
\usage{
\method{predict}{bl}(object, data, keep_rows = FALSE, val = "black_listed",
  ...)
}
\arguments{
\item{object}{a bl object ('blacklist')}

\item{data}{a data frame including factor variables}

\item{keep_rows}{logical, whether to keep all rows and fill with val}

\item{val}{charstring a value to fill the blacklisted objects with}

\item{...}{addtional params}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
