% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyscale.R
\name{predict.keyscale_obj}
\alias{predict.keyscale_obj}
\title{Predict Key Scale}
\usage{
\method{predict}{keyscale_obj}(object, data, unscale = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{object}{a keyscale_obj}

\item{data}{new data frame like object}

\item{unscale}{unscale a scaled object ( will auto read when scaled attr is set from previous predict)}

\item{verbose}{print debugging}

\item{...}{addtional params}
}
\value{
a data frame with either scale or unscaled values
}
\description{
creates seperate scales by a key
}
\author{
Matthew Davis
}
