% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_coder.R
\name{predict.level_coder_obj}
\alias{predict.level_coder_obj}
\title{predict method for level_coder_obj}
\usage{
\method{predict}{level_coder_obj}(object, data, rev = FALSE,
  return_primkey = FALSE, primkey_col = "primkey", verbose = FALSE, ...)
}
\arguments{
\item{object}{a 'level_coder_obj' class object}

\item{data}{data frame of new data}

\item{rev}{reverse the transformation}

\item{return_primkey}{use a primary key}

\item{primkey_col}{column to find the primary key}

\item{verbose}{print debugging}

\item{...}{addtional params}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
