% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_features}
\alias{ts_features}
\title{Time Series Features from Data Frame}
\usage{
ts_features(data, params = list(), verbose = TRUE)
}
\arguments{
\item{data}{a data frame or time series}

\item{verbose}{print debugging output}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
