% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_features}
\alias{ts_features}
\title{Time Series Features from Data Frame}
\usage{
ts_features(data, time_col = NULL, key_col = NULL, val_col = NULL,
  n_future = 6, n_past = 6)
}
\arguments{
\item{data}{a data frame or time series}

\item{time_col}{char time col in data}

\item{key_col}{char primary key for the data frame}

\item{val_col}{char value column in data frame}

\item{n_future}{int number of future steps}

\item{n_past}{int number of past steps (lags)}
}
\value{
a matrix of split text
}
\description{
splits a string in to chunks of length chunk_len, returns a matrix
}
\author{
Matthew Davis
}
